requirements.txt
PyQt6>=6.4.0
requests>=2.28.0
python-dotenv>=0.19.0
